# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: |
    SubNation Digital Subscriptions Platform API.

    ## Authentication (round-93 / 93-A8 F-1 — contract truth)

    Two session mechanisms, both documented as security schemes and
    attached per operation family:

    - **User routes** (auth, orders, wallet, cart, coupons/validate,
      loyalty, referrals, support, notifications): an `auth_token`
      httpOnly cookie (primary) OR an `Authorization: Bearer <token>`
      header (fallback).
    - **Admin routes** (/admin/*, /coupons/admin*, the copilot family):
      an `admin_token` httpOnly cookie (8h, issued by POST
      /admin/login) OR its Bearer equivalent. POST /admin/login itself
      is the credential endpoint and needs no prior session.

    ## CSRF gate (applies to every state-changing request)

    Every POST/PUT/PATCH/DELETE under /api — regardless of cookie or
    Bearer authentication — must carry an Origin (or Referer) header
    matching the deployment's allow-list (CSRF_ALLOWED_ORIGINS /
    APP_ORIGINS / APP_URL); otherwise the request is rejected with
    403 FORBIDDEN before the route runs. Requests with neither Origin
    nor Referer are also rejected. Exemptions:
    /api/auth/firebase/session, /api/auth/firebase/refresh, /api/cwv,
    /api/webhook/*, and /health. Non-browser clients must therefore
    send a browser-like Origin header on mutations.

    ## Deployment-conditional behavior

...
servers:
  - url: /api
...
security:
  - cookieAuth: []
    bearerAuth: []
...
paths:
...
  /loyalty:
    get:
      operationId: getLoyalty
      tags: [loyalty]
      summary: Loyalty summary — points, tier, referral stats, conversion rate
      responses:
        "200":
          description: Loyalty summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltySummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User row vanished (NOT_FOUND)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /loyalty/referrals:
    get:
      operationId: listReferrals
      tags: [loyalty]
      summary: List the current user's referral events (≤200, newest first)
      responses:
        "200":
          description: Referral events with masked referee phones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReferralEventItem"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Support ──────────────────────────────────────────────────────────────
...
  /admin/referrals/{id}/credit:
    post:
      operationId: creditReferral
      tags: [admin]
      security:
        - adminCookieAuth: []
          adminBearerAuth: []
      summary: Credit the referrer +50 loyalty points for a referral event (requireAdmin + users scope)
      description: |
        Status-guarded in a transaction (WHERE status='pending') — a
        concurrent credit loses the race and gets 409 ALREADY_EXISTS
        instead of double-granting. Mounted behind the idempotency
        middleware (routeKey admin.referrals.credit). A loyalty
        notification is created for the referrer.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: referral_events.id
        - name: Idempotency-Key
          in: header
          required: false
          description: |
            Idempotency key (≥ 8 chars to engage; 24h dedup per admin +
            route) — dedupes admin double-clicks.
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: Points granted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  points_credited:
                    type: integer
                    description: Points granted (POINTS_PER_REFERRAL = 50).
                required: [success, points_credited]
        "400":
          description: Event already credited (INVALID_DATA)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Missing users permission scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Referral event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Concurrent credit race (ALREADY_EXISTS)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Transaction failure (INTERNAL_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    # Round-93 (93-A8 F-1): the spec previously documented zero security
    # definitions while the middleware reality is cookie-first with a
    # Bearer fallback on user routes, and a separate admin cookie (8h)
    # plus Bearer on admin routes. Both alternatives are expressed per
    # operation via `security: [{ schemeA: [], schemeB: [] }]` (same
    # object = OR).
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: |
        User session — httpOnly cookie set by the auth routes (Google /
        Telegram / WhatsApp-OTP sign-in). SameSite=None in production so
        the Vercel-hosted SPA can call the API cross-site. Primary
        mechanism for browser clients.
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        User-session Bearer fallback — `Authorization: Bearer <token>`
        accepted wherever cookieAuth is (requireUser checks the cookie
        first, then the header).
    adminCookieAuth:
      type: apiKey
      in: cookie
      name: admin_token
      description: |
        Admin session — separate httpOnly cookie (8h) issued by POST
        /admin/login (and verify-2fa when TOTP is enabled).
    adminBearerAuth:
      type: http
      scheme: bearer
      description: |
        Admin-session Bearer alternative — the token returned in the
        POST /admin/login JSON body, accepted wherever adminCookieAuth
        is (requireAdmin accepts either).
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          $ref: "#/components/schemas/ErrorCode"
        details:
          type: object
          additionalProperties: true
          description: |
            Optional structured context — zod `issues` on validation
            failures, `{field}` hints on flash-sale validation errors,
            `{skipped_duplicates}` on inventory uploads.
      required: [error, code]

    # Machine-readable error codes — mirrors the shared ErrorCode enum
    # in @workspace/error-codes (shared/error-codes/src/index.ts), the
    # single source both backend and frontend re-export. The copilot
    # preview/execution family (round-92 B3/F-05) closes the gap where
    # backend/src/routes/admin/copilot/* returned codes outside the enum.
    ErrorCode:
      type: string
      enum:
        - INVALID_DATA
        - INVALID_PHONE
        - INVALID_PASSWORD_LENGTH
        - INVALID_PASSWORD_WEAK
        - INVALID_OTP
        - INVALID_CREDENTIAL
        - UNAUTHORIZED
        - INVALID_TOKEN
        - SESSION_EXPIRED
        - ACCOUNT_LOCKED
        - ACCOUNT_NOT_FOUND
        - PHONE_ALREADY_REGISTERED
        - FEATURE_DISABLED
        - FORBIDDEN
        - INSUFFICIENT_PERMISSIONS
        - NOT_FOUND
        - ALREADY_EXISTS
        - CONFLICT
        - OUT_OF_STOCK
        - PRODUCT_UNAVAILABLE
        - INSUFFICIENT_BALANCE
        - INVALID_AMOUNT
        - TOPUP_LIMIT_EXCEEDED
        - ORDER_NOT_FOUND
        - ORDER_ALREADY_COMPLETED
        - ORDER_CANNOT_CANCEL
        - GOOGLE_TOKEN_INVALID
        - GOOGLE_VERIFICATION_FAILED
        - INTERNAL_ERROR
        - SERVICE_UNAVAILABLE
        - RATE_LIMITED
        - IDEMPOTENCY_IN_FLIGHT
        - IDEMPOTENCY_KEY_REUSE
        - COPILOT_INVALID_INPUT
        - COPILOT_LLM_ERROR
        - COPILOT_NO_ADMIN_SESSION
        - COPILOT_BAD_METHOD
        - COPILOT_PREVIEW_NOT_FOUND
        - COPILOT_PREVIEW_CONSUMED
        - COPILOT_PREVIEW_EXPIRED
        - COPILOT_HANDOFF_REQUIRED
        - COPILOT_HIGH_RISK_DISABLED
        - COPILOT_STALE_RECORD
        - COPILOT_UNEXPECTED_STATE
        - COPILOT_EXECUTE_FAILED
        - COPILOT_NOT_HIGH_RISK
        - COPILOT_FIRST_CONFIRM_MISSING
        - COPILOT_COOLDOWN_NOT_ELAPSED
        - COPILOT_LLM_UNAVAILABLE
        - COPILOT_OUT_OF_SCOPE
        - COPILOT_INVALID_FLAGS
        # Round-93 (93-A8 F-8/F-15): two emitted codes that previously
        # lived outside the whole code contract — the copilot rate
        # limiter's 429 body and the CSRF gate's fail-closed
        # misconfiguration branch. The enum stays exhaustive over what
        # the backend actually emits.
        - COPILOT_RATE_LIMITED
        - CSRF_CONFIG
        # Round-94 (A5-05): the last two emitted-but-unenumerated codes —
        # the copilot phase-gate 503 (requireCopilotPhase) and the
        # outbound secret-scan 502 abort (ask/draft routes).
        - COPILOT_PHASE_DISABLED
        - COPILOT_SECRET_LEAK

    # 409 body returned by the idempotency middleware
    # (backend/src/middlewares/idempotency.ts): the same Idempotency-Key
    # re-sent with a different body (IDEMPOTENCY_KEY_REUSE) or the same
    # key while the original request is still executing
    # (IDEMPOTENCY_IN_FLIGHT). `message` is a legacy alias kept for older
    # clients — `error` is the canonical field.
...
    LoyaltySummary:
      type: object
      description: |
        Constants: POINTS_PER_LYD = 100, POINTS_PER_REFERRAL = 50; tiers
        silver=500 / gold=2000 / platinum=5000 lifetime spend. NOTE:
        points_value_lyd is a STRING (toFixed(2)).
      properties:
        points:
          type: integer
        tier:
          type: string
          enum: [bronze, silver, gold, platinum]
        lifetime_spend:
          type: number
        referral_code:
          type: string
          description: "Empty string when unset."
        referral_link:
          type: string
        referred_by:
          type: ["string", "null"]
        referrals_total:
          type: integer
        referrals_credited:
          type: integer
        referrals_pending:
          type: integer
        points_value_lyd:
          type: string
          description: points / 100, formatted with 2 decimals (e.g. "12.00").
        next_tier:
          type: ["object", "null"]
          description: null at platinum.
          properties:
            tier:
              type: string
            label:
              type: string
              description: Arabic tier label ("فضي" / "ذهبي" / "بلاتيني").
            remaining:
              type: number
              description: Lifetime spend still needed for the next tier.
          required: [tier, label, remaining]
        tier_thresholds:
          type: object
          properties:
            silver:
              type: number
            gold:
              type: number
            platinum:
              type: number
          required: [silver, gold, platinum]
        points_rate:
          type: object
          properties:
            points_per_referral:
              type: integer
            points_per_lyd:
              type: integer
          required: [points_per_referral, points_per_lyd]
      required:
        [
          points,
          tier,
          lifetime_spend,
          referral_code,
          referral_link,
          referrals_total,
          referrals_credited,
          referrals_pending,
          points_value_lyd,
          next_tier,
          tier_thresholds,
          points_rate,
        ]
...
    ReferralEventItem:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, credited]
        phone_masked:
          type: string
          description: Referee phone masked ("091****789").
        created_at:
          type: string
        credited_at:
          type: ["string", "null"]
        points_earned:
          type: integer
          description: 50 when credited, else 0.
      required: [id, status, phone_masked, created_at, credited_at, points_earned]

    # ── Support ──────────────────────────────────────────────────────────────
...